Update forms to support new relation
