+++ app/models/cruise.rb
@@ -1,6 +1,9 @@
 class Cruise < ActiveRecord::Base
+  include PrimaryObserver
+
   has_many :observations
   has_and_belongs_to_many :users
+  belongs_to :primary_observer, class_name: Person
 
   validates_presence_of :ship, :starts_at, :ends_at, :primary_observer, :objective
