+++ app/models/person.rb
@@ -2,6 +2,7 @@ class Person < ActiveRecord::Base
   has_many :person_observations
   has_many :observations, through: :person_observations
   has_many :comments
+  has_many :cruises
 
   validates_presence_of :name
